+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+- Added support for copying and pasting a task via 'y' and 'p'
+
+
 0.3.7
 -----
 - Converted UI to update widgets incrementally